Extract "hasParentTag" function in a helper file
