+++ lib/rules/require-button-type.js
@@ -1,19 +1,10 @@
 import { builders as b } from 'ember-template-recast';
 
-import { match } from '../helpers/node-matcher.js';
+import hasParentTag from '../helpers/has-parent-tag.js';
 import Rule from './_base.js';
 
 const ERROR_MESSAGE = 'All `<button>` elements should have a valid `type` attribute';
 
-function hasParentTag(path, tag) {
-  let parents = [...path.parents()];
-  let refParentNode = {
-    tag,
-    type: 'ElementNode',
-  };
-  let hasHeadElementInParentPath = parents.some((parent) => match(parent.node, refParentNode));
-  return Boolean(hasHeadElementInParentPath);
-}
 export default class RequireButtonType extends Rule {
   logNode({ node, message }) {
     return this.log({
